Fix bug when saving cgminer config via API
There are two parameters:
    '-c' sets path of config to load
    '--default-config' sets the path of config that's default

One would expect that if '-c' config name is passed, then a "save" command
would default to that path. This is not the case however, when "save" command
is invoked without arguments, a "--default-config" gets used.

This patch supplies the "--default-config" option set to the same path as "-c"
option.